+++ lib/steam_id.rb
@@ -14,7 +14,8 @@ module SteamID
   #   id1 = SteamID.from_string('STEAM_0:0:24110655')
   #   id2 = SteamID.from_string('gabenewell', api_key: '...')
   #   puts "Accounts are equal: #{ id1.account_id == id2.account_id }"
-  def from_string(s, api_key: nil)
-
+  def self.from_string(s, api_key: nil)
+    parser = Parser.new(api_key: api_key)
+    parser.from_string(s)
   end
 end