+++ CHANGELOG.md
@@ -12,6 +12,8 @@ glance - what to expact from upgrading to a new version.
 
 ### Fixed
 
+- Missing functionality of `SteamID.from_string`.
+
 ### Security
 
 ### Deprecated